notification-service/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0

opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-instrumentation-fastapi==0.45b0
